==> SNF_CNN/P01_snfpy.ipynb <==
from snf import datasets
import numpy as np
import pandas as pd
import snf
from snf import compute
import sklearn

digits = datasets.load_digits()
digits.keys()

for arr in digits.data:
    print(arr.shape)

groups, samples = np.unique(digits.labels, return_counts=True)

for grp, count in zip(groups, samples):
    print('Group {:.0f}: {} samples'.format(grp, count))

offsidesF = pd.read_csv('data/offsides_feature.csv', header=None)
structureF = pd.read_csv('data/structure_feature.csv', header=None)
print(offsidesF.shape)
# print(offsidesF.isnull().sum().sum())
structureF = np.array(structureF.values, dtype=np.float64)
offsidesF = np.array(offsidesF.values, dtype=np.float64)
structureFSim = sklearn.metrics.pairwise.cosine_similarity(structureF, Y=None, dense_output=True)
offsidesFSim = sklearn.metrics.pairwise.cosine_similarity(offsidesF, Y=None, dense_output=True)
data = [offsidesFSim, structureFSim]
data

# affinity_networks = compute.make_affinity(data, metric='cosine', K=9150, mu=0.5,normalize=True)
# # simcosine = pd.DataFrame(affinity_networks)
# # simcosine.to_csv("cosineSim.csv",index=False)

# len(affinity_networks[1])

# chem =  np.array(pd.read_csv('chem_Jacarrd_sim.csv', header=-1))
# enzyme =  np.array(pd.read_csv('enzyme_Jacarrd_sim.csv', header=-1))
# # chem = array(chem,enzyme)
# affinity_network = [chem,enzyme]

fused_network = snf.snf(data, K=20)

fused_network.shape

fused_network

# digits.data

fused_network = pd.DataFrame(fused_network)
fused_network.to_csv('CosineSNF(structureAndSideEffect).csv',index = False,header=False)
